pyproject.toml
[project]
name = "whoop-mcp"
version = "0.1.0"
description = "Personal MCP server wrapping the Whoop API v2 (workouts, sleep, recovery)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
whoop-mcp = "whoop_mcp.server:main"
whoop-mcp-login = "whoop_mcp.login:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whoop_mcp"]
